requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
imageio==2.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.3
# opencv-contrib-python==4.1.0.25
opencv-python==4.1.2.30
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pymdptoolbox==4.0b3
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
PyWavelets==1.0.3
pyzmq==18.0.1
qtconsole==4.5.1
scikit-image==0.15.0
scipy==1.1.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2


Flask
pandas
